refactor: move MultiProofTargetsV2 and ChunkedMultiProofTargetsV2 to reth-trie-common
Move these types from reth-trie-parallel into reth-trie-common's new
target_v2 module alongside ProofV2Target. Remove the targets_v2 module
from reth-trie-parallel. Update all call sites to import directly.

Amp-Thread-ID: https://ampcode.com/threads/T-019c9475-4446-7287-9cf3-a147508f942f
