+++ crates/trie/parallel/src/lib.rs
@@ -23,9 +23,6 @@ pub mod proof_task;
 /// Async value encoder for V2 proofs.
 pub(crate) mod value_encoder;
 
-/// V2 multiproof targets and chunking.
-pub mod targets_v2;
-
 /// Parallel state root metrics.
 #[cfg(feature = "metrics")]
 pub mod metrics;
